Extract the bot definition to separate file.
